@@ -8,7 +8,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0",
+        "php": ">=5.4.0",
         "illuminate/support": "4.*",
         "intervention/image": "2.*",
 		"doctrine/dbal": "2.*"
