@@ -2,10 +2,10 @@
 
 class ObjectController extends \BaseController {
 
-	private static $direction = array(
+	private static $direction = [
 		'asc'=>'Ascending',
 		'desc'=>'Descending',
-	);
+	];
 	
 	# Display list for home page
 	public function index() {
@@ -20,23 +20,23 @@ class ObjectController extends \BaseController {
 			$object->updated_by = $object->name;
 			if ($object->count == 0) $object->instance_count = '';
 		}
-		return View::make('avalon::objects.index', array('objects'=>$objects));
+		return View::make('avalon::objects.index', ['objects'=>$objects]);
 	}
 	
 	# Display create object form
 	public function create() {
-		$order_by = array(trans('avalon::messages.fields_system')=>array(
+		$order_by = [trans('avalon::messages.fields_system')=>[
 			'id'=>trans('avalon::messages.fields_id'),
 			'precedence'=>trans('avalon::messages.fields_precedence'),
 			'created_at'=>trans('avalon::messages.fields_created_at'),
 			'updated_at'=>trans('avalon::messages.fields_updated_at'),
-		));
+		]];
 
-		return View::make('avalon::objects.create', array(
-			'order_by'	=>$order_by,
+		return View::make('avalon::objects.create', [
+			'order_by' =>$order_by,
 			'direction'	=>self::$direction,
 			'list_groupings' =>self::getGroupings(),
-		));
+		]);
 	}
 	
 	# Store create object form post data
@@ -57,7 +57,7 @@ class ObjectController extends \BaseController {
 		if ($order_by == 'precedence') $direction = 'asc';
 
 		//create entry in objects table for new object
-		$object_id = DB::table(DB_OBJECTS)->insertGetId(array(
+		$object_id = DB::table(DB_OBJECTS)->insertGetId([
 			'title'			=> $title,
 			'name'			=> $name,
 			'model'			=> $model,
@@ -66,10 +66,10 @@ class ObjectController extends \BaseController {
 			'list_grouping'	=> Input::get('list_grouping'),
 			'updated_at'	=> new DateTime,
 			'updated_by'	=> Auth::user()->id,
-		));
+		]);
 		
 		//create title field for table by default
-		DB::table(DB_FIELDS)->insert(array(
+		DB::table(DB_FIELDS)->insert([
 			'title'			=> 'Title',
 			'name'			=> 'title',
 			'type'			=> 'string',
@@ -79,7 +79,7 @@ class ObjectController extends \BaseController {
 			'updated_at'	=> new DateTime,
 			'updated_by'	=> Auth::user()->id,
 			'precedence'	=> 1
-		));
+		]);
 		
 		//create table with boilerplate fields
 		Schema::create($name, function($table){
@@ -100,16 +100,16 @@ class ObjectController extends \BaseController {
 		//get order by select data
 		$object = DB::table(DB_OBJECTS)->where('name', $object_name)->first();
 		$fields = DB::table(DB_FIELDS)->where('object_id', $object->id)->orderBy('precedence')->get();
-		$order_by = array();
+		$order_by = [];
 		foreach ($fields as $field) $order_by[$field->name] = $field->title;
-		$order_by = array(
-			trans('avalon::messages.fields_system')=>array(
+		$order_by = [
+			trans('avalon::messages.fields_system')=>[
 				'id'=>trans('avalon::messages.fields_id'),
 				'precedence'=>trans('avalon::messages.fields_precedence'),
 				'created_at'=>trans('avalon::messages.fields_updated_at'),
-			),
+			],
 			trans('avalon::messages.fields_user')=>$order_by,
-		);
+		];
 
 		//related objects are different than dependencies; it's the subset of dependencies that are grouped by this object
 		$related_objects = DB::table(DB_FIELDS)
@@ -123,16 +123,16 @@ class ObjectController extends \BaseController {
 		$links = DB::table(DB_OBJECT_LINKS)->where('object_id', $object->id)->lists('linked_id');
 
 		//return view
-		return View::make('avalon::objects.edit', array(
-			'object'			=>DB::table(DB_OBJECTS)->where('id', $object->id)->first(), 
+		return View::make('avalon::objects.edit', [
+			'object'			=>$object, 
 			'order_by'			=>$order_by,
 			'direction'			=>self::$direction,
 			'dependencies'		=>DB::table(DB_FIELDS)->where('related_object_id', $object->id)->count(),
-			'group_by_field'	=>array(''=>'') + DB::table(DB_FIELDS)->where('object_id', $object->id)->where('type', 'select')->lists('title', 'id'),
+			'group_by_field'	=>[''=>''] + DB::table(DB_FIELDS)->where('object_id', $object->id)->where('type', 'select')->lists('title', 'id'),
 			'list_groupings'	=>self::getGroupings(),
 			'related_objects'	=>$related_objects,
 			'links'				=>$links,
-		));
+		]);
 	}
 	
 	//edit object settings
@@ -151,33 +151,32 @@ class ObjectController extends \BaseController {
 		//not sure why it's necessary, doesn't like empty value all of a sudden
 		$group_by_field = Input::has('group_by_field') ? Input::get('group_by_field') : null;
 
-		//is a singleton instance?
-		$singleton = Input::has('singleton') ? 1 : 0;
-
 		//linked objects
 		DB::table(DB_OBJECT_LINKS)->where('object_id', $object->id)->delete();
 		if (Input::has('related_objects')) {
 			foreach (Input::get('related_objects') as $linked_id) {
-				DB::table(DB_OBJECT_LINKS)->insert(array(
+				DB::table(DB_OBJECT_LINKS)->insert([
 					'object_id'=>$object->id,
 					'linked_id'=>$linked_id,
-				));
+				]);
 			}
 		}
 
 		//update objects table
-		DB::table(DB_OBJECTS)->where('id', $object->id)->update(array(
-			'title'				=>Input::get('title'),
-			'name'				=>$new_name,
-			'model'				=>Input::get('model'),
-			'order_by'			=>$order_by,
-			'direction'			=>$direction,
-			'singleton'			=>$singleton,
-			'list_grouping'		=>Input::get('list_grouping'),
-			'group_by_field'	=>$group_by_field,
-			'list_help'			=>trim(Input::get('list_help')),
-			'form_help'			=>trim(Input::get('form_help')),
-		));
+		DB::table(DB_OBJECTS)->where('id', $object->id)->update([
+			'title'				=> Input::get('title'),
+			'name'				=> $new_name,
+			'model'				=> Input::get('model'),
+			'order_by'			=> $order_by,
+			'direction'			=> $direction,
+			'singleton'			=> Input::has('singleton') ? 1 : 0,
+			'can_create'		=> Input::has('can_create') ? 1 : 0,
+			'can_edit'			=> Input::has('can_edit') ? 1 : 0,
+			'list_grouping'		=> Input::get('list_grouping'),
+			'group_by_field'	=> $group_by_field,
+			'list_help'			=> trim(Input::get('list_help')),
+			'form_help'			=> trim(Input::get('form_help')),
+		]);
 		
 		return Redirect::action('InstanceController@index', $new_name);
 	}
@@ -205,13 +204,13 @@ class ObjectController extends \BaseController {
 		$tables = array_diff($tables, DB::table(DB_FIELDS)->where('type', 'checkboxes')->lists('name'));
 
 		//remove the avalon system tables
-		$tables = array_diff($tables, array(DB_FIELDS, DB_FILES, DB_OBJECT_LINKS, DB_OBJECT_USER, DB_OBJECTS, DB_USERS));
+		$tables = array_diff($tables, [DB_FIELDS, DB_FILES, DB_OBJECT_LINKS, DB_OBJECT_USER, DB_OBJECTS, DB_USERS]);
 
 		//remove avalon paths
-		$tables = array_diff($tables, array('create', 'logout'));
+		$tables = array_diff($tables, ['create', 'logout']);
 		
 		//remove the laravel system tables
-		$tables = array_diff($tables, array('migrations'));
+		$tables = array_diff($tables, ['migrations']);
 
 		dd($tables);
 	}
@@ -225,7 +224,7 @@ class ObjectController extends \BaseController {
 
 	//get array of tables in database
 	private static function getTables() {
-		$return = array();
+		$return = [];
 		$pdo = DB::connection()->getPdo();
 		$tables = $pdo->query('SHOW TABLES');
 		foreach ($tables as $table) $return[] = $table[0];
