@@ -1,18 +1,18 @@
 @extends('avalon::template')
 
 @section('title')
-	{{ Lang::get('avalon::messages.instances_edit') }}
+	@lang('avalon::messages.instances_edit')
 @endsection
 
 @section('main')
 
-	{{ Breadcrumbs::leave(array(
-		URL::action('ObjectController@index')=>Lang::get('avalon::messages.objects'),
+	{{ Breadcrumbs::leave([
+		URL::action('ObjectController@index')=>trans('avalon::messages.objects'),
 		URL::action('InstanceController@index', $object->name)=>$object->title,
-		Lang::get('avalon::messages.instances_edit'),
-		)) }}
+		trans('avalon::messages.instances_edit'),
+		]) }}
 
-	{{ Form::open(array('class'=>'form-horizontal ' . $object->name, 'url'=>URL::action('InstanceController@update', array($object->name, $instance->id, $linked_id)), 'method'=>'put')) }}
+	{{ Form::open(['class'=>'form-horizontal ' . $object->name, 'url'=>URL::action('InstanceController@update', [$object->name, $instance->id, $linked_id]), 'method'=>'put']) }}
 	
 	@if (Input::has('return_to'))
 		{{ Form::hidden('return_to', Input::get('return_to')) }}
@@ -37,7 +37,7 @@
 					</label>
 					@endforeach
 				@elseif ($field->type == 'color')
-					{{ Form::text($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ' {hash:true,caps:false}' . ($field->required ? ' required' : ''))) }}
+					{{ Form::text($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ' {hash:true,caps:false}' . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'date')
 					<div class="input-group date" data-date-format="MM/DD/YYYY">
 						<span class="input-group-addon"><span class="glyphicon glyphicon-calendar"></span></span>
@@ -49,7 +49,7 @@
 						<input type="text" class="form-control  @if ($field->required) required @endif" value="{{ $instance->{$field->name} }}" name="{{ $field->name }}">
 					</div>
 				@elseif ($field->type == 'html')
-					{{ Form::textarea($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::textarea($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'image')
 					@if (empty($instance->{$field->name}))
 					<div class="image new" data-field-id="{{ $field->id }}" style="width:{{ $field->screen_width }}px; height:{{ $field->screen_height }}px; line-height:{{ $field->screen_height }}px;">
@@ -63,7 +63,7 @@
 					{{ Form::hidden($field->name, $instance->{$field->name}->id) }}
 					@endif
 				@elseif ($field->type == 'images')
-					<?php $ids = array(); ?>
+					<?php $ids = []; ?>
 					@foreach ($instance->{$field->name} as $image)
 						<div class="image" data-file-id="{{ $image->id }}" data-field-id="{{ $field->id }}" style="width:{{ $field->screen_width }}px; height:{{ $field->screen_height }}px; line-height:{{ $field->screen_height }}px; background-image: url({{ $image->url }});">
 							{{ $field->width }} &times; {{ $field->height }}
@@ -75,22 +75,22 @@
 						{{ $field->width or '&infin;' }} &times; {{ $field->height or '&infin;' }}
 					</div>
 				@elseif ($field->type == 'integer')
-					{{ Form::integer($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::integer($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'select')
-					{{ Form::select($field->name, $field->options, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::select($field->name, $field->options, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'slug')
-					{{ Form::text($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::text($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'string')
-					{{ Form::text($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::text($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'text')
-					{{ Form::textarea($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::textarea($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@elseif ($field->type == 'time')
 					<div class="input-group time" data-date-format="hh:mm A">
 						<span class="input-group-addon"><span class="glyphicon glyphicon-time"></span></span>
-						{{ Form::text($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::text($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					</div>
 				@elseif ($field->type == 'url')
-					{{ Form::url($field->name, $instance->{$field->name}, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+					{{ Form::url($field->name, $instance->{$field->name}, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 				@endif
 			</div>
 		</div>
@@ -98,11 +98,11 @@
 	
 	<div class="form-group">
 		<div class="col-sm-10 col-sm-offset-2">
-			{{ Form::submit(Lang::get('avalon::messages.site_save'), array('class'=>'btn btn-primary')) }}
+			{{ Form::submit(trans('avalon::messages.site_save'), ['class'=>'btn btn-primary']) }}
 			@if (Input::has('return_to'))
-			{{ HTML::link(Input::get('return_to'), Lang::get('avalon::messages.site_cancel'), array('class'=>'btn btn-default')) }}
+			{{ HTML::link(Input::get('return_to'), trans('avalon::messages.site_cancel'), ['class'=>'btn btn-default']) }}
 			@else
-			{{ HTML::link(URL::action('InstanceController@index', $object->name), Lang::get('avalon::messages.site_cancel'), array('class'=>'btn btn-default')) }}
+			{{ HTML::link(URL::action('InstanceController@index', $object->name), trans('avalon::messages.site_cancel'), ['class'=>'btn btn-default']) }}
 			@endif
 		</div>
 	</div>
@@ -115,7 +115,10 @@
 		<h3>{{ $link['object']->title }}</h3>
 
 		<div class="btn-group">
-			<a class="btn btn-default" id="create" href="{{ URL::action('InstanceController@create', array($link['object']->name, $instance->id)) }}"><i class="glyphicon glyphicon-plus"></i> {{ Lang::get('avalon::messages.instances_create') }}</a>
+			<a class="btn btn-default" id="create" href="{{ URL::action('InstanceController@create', [$link['object']->name, $instance->id]) }}">
+				<i class="glyphicon glyphicon-plus"></i> 
+				@lang('avalon::messages.instances_create')
+			</a>
 		</div>
 		
 		{{ InstanceController::table($link['object'], $link['fields'], $link['instances']) }}
@@ -130,8 +133,8 @@
 		<p>{{ nl2br($object->form_help) }}</p>
 	@endif
 
-	{{ Form::open(array('method'=>'delete', 'action'=>array('InstanceController@destroy', $object->name, $instance->id))) }}
-	<button type="submit" class="btn btn-default btn-xs">{{ Lang::get('avalon::messages.instances_destroy') }}</button>
+	{{ Form::open(['method'=>'delete', 'action'=>['InstanceController@destroy', $object->name, $instance->id]]) }}
+	<button type="submit" class="btn btn-default btn-xs">@lang('avalon::messages.instances_destroy')</button>
 	{{ Form::close() }}
 
 @endsection
