@@ -1,3 +1,5 @@
-<p>{{ Lang::get('avalon::messages.users_welcome_message', array('link'=>$link)) }}</p>
-<p>{{ Lang::get('avalon::messages.users_email') }} : {{ $email }}<br>
-{{ Lang::get('avalon::messages.users_password') }} : {{ $password }}</p>
+<p>@lang('avalon::messages.users_welcome_message', ['link'=>$link])</p>
+<p>
+	@lang('avalon::messages.users_email') : {{ $email }}<br>
+	@lang('avalon::messages.users_password') : {{ $password }}
+</p>
