@@ -1,18 +1,18 @@
 @extends('avalon::template')
 
 @section('title')
-	{{ Lang::get('avalon::messages.instances_create') }}
+	@lang('avalon::messages.instances_create')
 @endsection
 
 @section('main')
 
-	{{ Breadcrumbs::leave(array(
-		URL::action('ObjectController@index')=>Lang::get('avalon::messages.objects'),
+	{{ Breadcrumbs::leave([
+		URL::action('ObjectController@index')=>trans('avalon::messages.objects'),
 		URL::action('InstanceController@index', $object->name)=>$object->title,
-		Lang::get('avalon::messages.instances_create'),
-		)) }}
+		trans('avalon::messages.instances_create'),
+		]) }}
 
-	{{ Form::open(array('class'=>'form-horizontal', 'url'=>URL::action('InstanceController@store', array($object->name, $linked_id)))) }}
+	{{ Form::open(['class'=>'form-horizontal', 'url'=>URL::action('InstanceController@store', [$object->name, $linked_id])]) }}
 	
 	@if (Input::has('return_to'))
 		{{ Form::hidden('return_to', Input::get('return_to')) }}
@@ -35,7 +35,7 @@
 							</label>
 						@endforeach
 					@elseif ($field->type == 'color')
-						{{ Form::text($field->name, $field->required ? '#ffffff' : null, array('class'=>'form-control ' . $field->type . ' {hash:true,caps:false}' . ($field->required ? ' required' : ''))) }}
+						{{ Form::text($field->name, $field->required ? '#ffffff' : null, ['class'=>'form-control ' . $field->type . ' {hash:true,caps:false}' . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'date')
 						<div class="input-group date" data-date-format="MM/DD/YYYY">
 							<span class="input-group-addon"><span class="glyphicon glyphicon-calendar"></span></span>
@@ -55,7 +55,7 @@
 						   	@endif
 						</div>
 					@elseif ($field->type == 'html')
-						{{ Form::textarea($field->name, null, array('class'=>'form-control html' . ($field->required ? ' required' : ''))) }}
+						{{ Form::textarea($field->name, null, ['class'=>'form-control html' . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'image')
 						{{ Form::hidden($field->name, null) }}
 						<div class="image new" data-field-id="{{ $field->id }}" style="width:{{ $field->screen_width }}px; height:{{ $field->screen_height }}px; line-height:{{ $field->screen_height }}px;">
@@ -67,22 +67,22 @@
 							<span>{{ $field->width or '&infin;' }} &times; {{ $field->height or '&infin;' }}</span>
 						</div>
 					@elseif ($field->type == 'integer')
-						{{ Form::integer($field->name, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::integer($field->name, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'select')
-						{{ Form::select($field->name, $field->options, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::select($field->name, $field->options, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'slug')
-						{{ Form::text($field->name, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::text($field->name, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'string')
-						{{ Form::text($field->name, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::text($field->name, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'text')
-						{{ Form::textarea($field->name, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::textarea($field->name, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					@elseif ($field->type == 'time')
 						<div class="input-group time" data-date-format="hh:mm A">
 							<span class="input-group-addon"><span class="glyphicon glyphicon-time"></span></span>
-							{{ Form::text($field->name, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+							{{ Form::text($field->name, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 						</div>
 					@elseif ($field->type == 'url')
-						{{ Form::url($field->name, null, array('class'=>'form-control ' . $field->type . ($field->required ? ' required' : ''))) }}
+						{{ Form::url($field->name, null, ['class'=>'form-control ' . $field->type . ($field->required ? ' required' : '')]) }}
 					@endif
 				</div>
 			</div>
@@ -91,8 +91,8 @@
 	
 	<div class="form-group">
 		<div class="col-sm-10 col-sm-offset-2">
-			{{ Form::submit(Lang::get('avalon::messages.site_save'), array('class'=>'btn btn-primary')) }}
-			{{ HTML::link(URL::action('InstanceController@index', $object->name), Lang::get('avalon::messages.site_cancel'), array('class'=>'btn btn-default')) }}
+			{{ Form::submit(trans('avalon::messages.site_save'), ['class'=>'btn btn-primary']) }}
+			{{ HTML::link(URL::action('InstanceController@index', $object->name), trans('avalon::messages.site_cancel'), ['class'=>'btn btn-default']) }}
 		</div>
 	</div>
 
