@@ -6,17 +6,28 @@
 
 @section('main')
 
-	{{ Breadcrumbs::leave(array(
-		URL::action('ObjectController@index')=>Lang::get('avalon::messages.objects'),
+	{{ Breadcrumbs::leave([
+		URL::action('ObjectController@index')=>trans('avalon::messages.objects'),
 		$object->title,
-		)) }}
+		]) }}
 
 	<div class="btn-group">
 		@if (Auth::user()->role < 2)
-		<a class="btn btn-default" href="{{ URL::action('ObjectController@edit', $object->name) }}"><i class="glyphicon glyphicon-cog"></i> {{ Lang::get('avalon::messages.objects_edit', array('title'=>$object->title)) }}</a>
-		<a class="btn btn-default" href="{{ URL::action('FieldController@index', $object->name) }}"><i class="glyphicon glyphicon-list"></i> {{ Lang::get('avalon::messages.fields') }}</a>
+		<a class="btn btn-default" href="{{ URL::action('ObjectController@edit', $object->name) }}">
+			<i class="glyphicon glyphicon-cog"></i> 
+			@lang('avalon::messages.objects_edit', ['title'=>$object->title])
+		</a>
+		<a class="btn btn-default" href="{{ URL::action('FieldController@index', $object->name) }}">
+			<i class="glyphicon glyphicon-list"></i>
+			@lang('avalon::messages.fields')
+		</a>
 		@endif
-		<a class="btn btn-default" id="create" href="{{ URL::action('InstanceController@create', $object->name) }}"><i class="glyphicon glyphicon-plus"></i> {{ Lang::get('avalon::messages.instances_create') }}</a>
+		@if ($object->can_create)
+			<a class="btn btn-default" id="create" href="{{ URL::action('InstanceController@create', $object->name) }}">
+				<i class="glyphicon glyphicon-plus"></i>
+				@lang('avalon::messages.instances_create')
+			</a>
+		@endif
 	</div>
 
 	@if (count($instances))
@@ -26,14 +37,14 @@
 					Title
 					<div class="updated_at">Updated</div>
 				</div>
-				@include('avalon::instances.nested', array('instances'=>$instances))
+				@include('avalon::instances.nested', ['instances'=>$instances])
 			</div>
 		@else
 			{{ InstanceController::table($object, $fields, $instances) }}
 		@endif
 	@else
 	<div class="alert alert-warning">
-		{{ Lang::get('avalon::messages.instances_empty', array('title'=>strtolower($object->title))) }}
+		@lang('avalon::messages.instances_empty', ['title'=>strtolower($object->title)])
 	</div>
 	@endif
 
