@@ -1,41 +1,41 @@
 @extends('avalon::template')
 
 @section('title')
-	{{ Lang::get('avalon::messages.users_create') }}
+	@lang('avalon::messages.users_create')
 @endsection
 
 @section('main')
 
-	{{ Breadcrumbs::leave(array(
-		URL::action('ObjectController@index')=>Lang::get('avalon::messages.objects'),
-		URL::action('UserController@index')=>Lang::get('avalon::messages.users'),
-		Lang::get('avalon::messages.users_create'),
-		)) }}
+	{{ Breadcrumbs::leave([
+		URL::action('ObjectController@index')=>trans('avalon::messages.objects'),
+		URL::action('UserController@index')=>trans('avalon::messages.users'),
+		trans('avalon::messages.users_create'),
+		]) }}
 
-	{{ Form::open(array('class'=>'form-horizontal', 'url'=>URL::action('UserController@store'))) }}
+	{{ Form::open(['class'=>'form-horizontal', 'url'=>URL::action('UserController@store')]) }}
 	
 	<div class="form-group">
-		{{ Form::label('name', Lang::get('avalon::messages.users_name'), array('class'=>'control-label col-sm-2')) }}
+		{{ Form::label('name', trans('avalon::messages.users_name'), ['class'=>'control-label col-sm-2']) }}
 	    <div class="col-sm-10">
-			{{ Form::text('name', false, array('class'=>'form-control required')) }}
+			{{ Form::text('name', false, ['class'=>'form-control required']) }}
 	    </div>
 	</div>
 
 	<div class="form-group">
-		{{ Form::label('email', Lang::get('avalon::messages.users_email'), array('class'=>'control-label col-sm-2')) }}
+		{{ Form::label('email', trans('avalon::messages.users_email'), ['class'=>'control-label col-sm-2']) }}
 	    <div class="col-sm-10">
-			{{ Form::email('email', false, array('class'=>'form-control required')) }}
+			{{ Form::email('email', false, ['class'=>'form-control required']) }}
 	    </div>
 	</div>
 	
 	<div class="form-group">
-		{{ Form::label('role', Lang::get('avalon::messages.users_role'), array('class'=>'control-label col-sm-2')) }}
+		{{ Form::label('role', trans('avalon::messages.users_role'), ['class'=>'control-label col-sm-2']) }}
 	    <div class="col-sm-10">
 	    	@foreach ($roles as $role_id=>$role)
 			<div class="radio">
 				<label>
 					{{ Form::radio('role', $role_id, $role_id == 3) }}
-					<strong>{{ $role }}</strong> &middot; {{ Lang::get('avalon::messages.users_role_' . Str::slug($role)) }}
+					<strong>{{ $role }}</strong> &middot; {{ trans('avalon::messages.users_role_' . Str::slug($role)) }}
 				</label>
 			</div>
 			@endforeach
@@ -44,8 +44,8 @@
 
 	<div class="form-group">
 	    <div class="col-sm-10 col-sm-offset-2">
-			{{ Form::submit(Lang::get('avalon::messages.site_save'), array('class'=>'btn btn-primary')) }}
-			{{ HTML::link(URL::action('UserController@index'), Lang::get('avalon::messages.site_cancel'), array('class'=>'btn btn-default')) }}
+			{{ Form::submit(trans('avalon::messages.site_save'), ['class'=>'btn btn-primary']) }}
+			{{ HTML::link(URL::action('UserController@index'), trans('avalon::messages.site_cancel'), ['class'=>'btn btn-default']) }}
 	    </div>
 	</div>
 	
@@ -54,5 +54,5 @@
 @endsection
 
 @section('side')
-	<p>{{ Lang::get('avalon::messages.users_form_help') }}</p>
+	<p>@lang('avalon::messages.users_form_help')</p>
 @endsection
